pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "ccxt>=4.4.92",
    "flask>=3.1.1",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
